Removed previous font implementation
Deleted FontSTB in favor of FontDetails and TextRenderer.
Renamed FontLoaderSTB to no longer specifically only reference STB for
now.
Moved font resources into proper resource folder.
