+++ base/src/main/java/com/avogine/render/data/FontDetails.java
@@ -1,5 +1,6 @@
 package com.avogine.render.data;
 
+import org.lwjgl.opengl.GL11;
 import org.lwjgl.stb.*;
 
 /**
@@ -12,4 +13,13 @@ import org.lwjgl.stb.*;
  */
 public record FontDetails(int fontSize, float descent, float scale, STBTTFontinfo info, STBTTPackedchar.Buffer cdata, int textureID) {
 
+	/**
+	 * Free up any memory in use by this {@link FontDetails}.
+	 */
+	public void cleanup() {
+		GL11.glDeleteTextures(textureID);
+		info.free();
+		cdata.free();
+	}
+	
 }
